+++ src/prime_rl/inference/vllm/worker/nixl.py
@@ -19,12 +19,12 @@ from vllm.config import set_current_vllm_config
 from vllm.logger import init_logger
 
 from prime_rl.inference.vllm.worker.weight_transfer import update_mla_absorbed_weights
-from prime_rl.trainer.rl.broadcast.nixl.agent import MemDesc, NixlAgent, make_agent_name, set_ucx_env_defaults
-from prime_rl.trainer.rl.broadcast.nixl.cuda_malloc_memory import (
+from prime_rl.transports.weights.nixl.agent import MemDesc, NixlAgent, make_agent_name, set_ucx_env_defaults
+from prime_rl.transports.weights.nixl.cuda_malloc_memory import (
     size_cuda_buffers,
     use_cuda_malloc_pool,
 )
-from prime_rl.trainer.rl.broadcast.nixl.graph import (
+from prime_rl.transports.weights.nixl.graph import (
     Destination,
     OperationChain,
     RecordedCopy,
@@ -34,9 +34,9 @@ from prime_rl.trainer.rl.broadcast.nixl.graph import (
     make_hf_lazy_weights,
     plan_tensor_replay,
 )
-from prime_rl.trainer.rl.broadcast.nixl.model_express import ModelExpressSession
-from prime_rl.trainer.rl.broadcast.nixl.tensor_routing import route_sharded_tensor
-from prime_rl.trainer.rl.broadcast.nixl.trainer_tensor_table import TrainerTensorTable
+from prime_rl.transports.weights.nixl.model_express import ModelExpressSession
+from prime_rl.transports.weights.nixl.tensor_routing import route_sharded_tensor
+from prime_rl.transports.weights.nixl.trainer_tensor_table import TrainerTensorTable
 
 if TYPE_CHECKING:
     from vllm.v1.worker.gpu_worker import Worker
