+++ src/prime_rl/orchestrator/algo/opsd.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from renderers.base import Renderer
 
     from prime_rl.orchestrator.types import Rollout
-    from prime_rl.transport import TrainingSample
+    from prime_rl.transports.rollouts import TrainingSample
     from prime_rl.utils.client import InferencePool
