+++ src/prime_rl/orchestrator/watcher.py
@@ -11,7 +11,7 @@ from modelexpress import p2p_pb2
 
 from prime_rl.configs.orchestrator import OrchestratorConfig
 from prime_rl.orchestrator.types import Policy, VersionObserver
-from prime_rl.trainer.rl.broadcast.nixl.model_express import ModelExpressSession
+from prime_rl.transports.weights.nixl.model_express import ModelExpressSession
 from prime_rl.utils.async_utils import safe_cancel
 from prime_rl.utils.client import InferencePool
 from prime_rl.utils.logger import format_time, get_logger
