+++ src/prime_rl/transports/weights/__init__.py
@@ -4,10 +4,10 @@ import torch
 
 from prime_rl.configs.trainer import LoRAConfig, WeightBroadcastConfig
 from prime_rl.trainer.parallel_dims import ParallelDims
-from prime_rl.trainer.rl.broadcast.base import WeightBroadcast
-from prime_rl.trainer.rl.broadcast.filesystem import FileSystemWeightBroadcast
-from prime_rl.trainer.rl.broadcast.nccl import NCCLWeightBroadcast
-from prime_rl.trainer.rl.broadcast.nixl import NIXLWeightBroadcast
+from prime_rl.transports.weights.base import WeightBroadcast
+from prime_rl.transports.weights.filesystem import FileSystemWeightBroadcast
+from prime_rl.transports.weights.nccl import NCCLWeightBroadcast
+from prime_rl.transports.weights.nixl import NIXLWeightBroadcast
 
 
 def setup_weight_broadcast(
