+++ src/prime_rl/transports/weights/nccl.py
@@ -14,8 +14,8 @@ from vllm.distributed.utils import StatelessProcessGroup
 from prime_rl.configs.trainer import NCCLWeightBroadcastConfig
 from prime_rl.trainer.conversion_utils import get_max_layer_num
 from prime_rl.trainer.models import PreTrainedModelPrimeRL
-from prime_rl.trainer.rl.broadcast.base import WeightBroadcast
 from prime_rl.trainer.utils import get_world
+from prime_rl.transports.weights.base import WeightBroadcast
 from prime_rl.utils.client import NCCL_READY_MARKER
 from prime_rl.utils.logger import get_logger
 from prime_rl.utils.nccl import disable_nccl_p2p_if_unavailable
