+++ src/prime_rl/transports/weights/nixl/__init__.py
@@ -1,5 +1,5 @@
 """NIXL weight broadcasting and transfer helpers."""
 
-from prime_rl.trainer.rl.broadcast.nixl.nixl import NIXLWeightBroadcast
+from prime_rl.transports.weights.nixl.nixl import NIXLWeightBroadcast
 
 __all__ = ["NIXLWeightBroadcast"]
