+++ src/prime_rl/transports/weights/nixl/tensor_routing.py
@@ -6,8 +6,8 @@ from dataclasses import dataclass
 
 import torch
 
-from prime_rl.trainer.rl.broadcast.nixl.graph import TensorReplayPlan
-from prime_rl.trainer.rl.broadcast.nixl.trainer_tensor_table import TrainerTensor
+from prime_rl.transports.weights.nixl.graph import TensorReplayPlan
+from prime_rl.transports.weights.nixl.trainer_tensor_table import TrainerTensor
 
 
 @dataclass(frozen=True)
