+++ tests/unit/transports/test_nccl_broadcast.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 
 from prime_rl.inference.vllm.worker.nccl import NCCLWeightBroadcastReceiver
-from prime_rl.trainer.rl.broadcast.nccl import NCCLWeightBroadcastSender
+from prime_rl.transports.weights.nccl import NCCLWeightBroadcastSender
 
 pytestmark = [pytest.mark.gpu]
